+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 Changes in dev
 ==============
 
+- Fixed Handle None values for non-required fields.
 - Removed Document._get_subclasses() - no longer required
 - Fixed bug requiring subclasses when not actually needed
 - Fixed deletion of dynamic data